+++ src/app/(auth)/reset-password/page.tsx
@@ -1,0 +1,26 @@
+"use client";
+
+import { useEffect } from "react";
+import { useRouter, useSearchParams } from "next/navigation";
+
+export default function ResetPasswordRedirect() {
+  const router = useRouter();
+  const searchParams = useSearchParams();
+
+  useEffect(() => {
+    const uid = searchParams.get("uid");
+    const token = searchParams.get("token");
+
+    if (uid && token) {
+      router.push(`/password/reset?uid=${uid}&token=${token}`);
+    } else {
+      router.push("/password/forgot");
+    }
+  }, [router, searchParams]);
+
+  return (
+    <div className="flex items-center justify-center min-h-screen bg-gray-50 dark:bg-gray-900">
+      <p className="text-gray-600 dark:text-gray-400">Redirecting...</p>
+    </div>
+  );
+}
